ajustes varios
