+++ sx-server/grails-app/conf/spring/resources.groovy
@@ -1,3 +1,7 @@
 // Place your Spring DSL code here
+import org.springframework.web.servlet.i18n.SessionLocaleResolver;
 beans = {
+	localeResolver(SessionLocaleResolver) { 
+        defaultLocale= new java.util.Locale('us'); 
+    }
 }
